CN-20

+++ database/migrations/2024_05_09_173234_create_cards_table.php
@@ -26,7 +26,7 @@ class CreateCardsTable extends Migration
             $table->string('account_name');
             $table->string('note')->nullable();
 
-            $table->smallInteger('fee_percent');
+            $table->integer('fee_percent');
             $table->bigInteger('total_money');
             $table->enum('formality', ['R', 'D']);
             // $table->bigInteger('fee');
